Changed position to i32 FOR NOW. Implemented basic player controls. Need to move to actual System architecture

+++ Components/src/Physics/mod.rs
@@ -4,7 +4,7 @@ use ggez::nalgebra::Point2;
 
 #[derive(Component)]
 pub struct Physics {
-    pub position: Point2<f32>,
+    pub position: Point2<i32>,
     pub velocity: i32,
     pub acceleration: i32
 }
